Removed some unneeded casts.
git-svn-id: file:///home/nick/svnrepo/repo/indigo/trunk@15769 20aa11e4-6e6f-594c-a032-6291ace9ee65

+++ networking/URL.cpp
@@ -16,14 +16,14 @@ URL URL::parseURL(const std::string& url) // throws glare::Exception
 	URL result;
 	result.port = -1;
 
-	Parser parser(url.data(), (unsigned int)url.size());
+	Parser parser(url.data(), url.size());
 
 	//---------------- Parse scheme ----------------
 	const size_t scheme_terminator_pos = url.find("://");
 	if(scheme_terminator_pos != std::string::npos)
 	{
 		result.scheme = url.substr(0, scheme_terminator_pos);
-		parser.setCurrentPos((unsigned int)scheme_terminator_pos + 3);
+		parser.setCurrentPos(scheme_terminator_pos + 3);
 	}
 
 	//---------------- Parse host ----------------
